Pathfinding::calculate()
- refactor that bit about strafe/ dash/ pathPreview
- and conform it to BattlescapeGame::primaryAction()

hopefully _battleAction(NULL) doesn't slip under the radar and CTD somewhere.

+++ src/Ruleset/RuleArmor.cpp
@@ -314,7 +314,7 @@ int RuleArmor::getSize() const
  */
 float RuleArmor::getDamageModifier(ItemDamageType dType) const
 {
-	return _damageModifier[static_cast<int>(dType)];
+	return _damageModifier[static_cast<size_t>(dType)];
 }
 
 /** Gets the Line Of Fire Template set.
